Add : Integrating with add category api

+++ app/src/main/java/co/id/billyon/model/CategoryRequest.kt
@@ -6,5 +6,5 @@ import com.google.gson.annotations.SerializedName
 data class CategoryRequest(
         @SerializedName("category_name") val categoryName: String,
         @SerializedName("image_path") val imagePath: String,
-        @SerializedName("store_id") val storeId: Int
+        @SerializedName("store_id") val storeId: Long
 )
